+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Next Release
 
 * Added `Mongoid::Shell::Commands::Mongo` that generates a command line which connects to the session's primary node - [@dblock](https://github.com/dblock).
 * A `Mongoid::Shell::Errors::SessionNotConnectedError` error will be raised when a session is not connected - [@dblock](https://github.com/dblock).
+* Added support for `--out` to `Mongoid::Shell::Commands::Mongodump` - [@dblock](https://github.com/dblock).
 
 0.1.0 (1/27/2013)
 ==================